RAG/requirements.txt
fastapi[all]
openai
tenacity
google-cloud-datastore
google-cloud-storage
azure-ai-formrecognizer
azure-core
tiktoken
pinecone